requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
anthropic==0.34.0
openai==1.45.0
python-dotenv==1.0.1
pydantic==2.9.0
